fs: Add destroy_inode callbacks to filesystems
Several filesystems rely on the default function which frees
the struct inode * rather than the filesystem specific inode
which the inode is embedded in. This works because the inode
is the first element in the filesystem specific inode. Let's
not depend on this behaviour and for clarity add the destroy_inode
callbacks to all filesystems.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>