mci: imx-esdhc-pbl: Add imx8mp_esdhc_load_image() for i.MX8MP
The image format of the i.MX8MP is different from i.MX8M, so add its own
image loading function for it.

Older i.MX SoCs had a IVT Offset (the offset from the start of the image
to the actual data) of 1KiB. This was done to leave space for the
partition table at the beginning of the device. To support GPT SoCs
starting with i.MX8M an additional gap of 32KiB was added, so that the
actual image started at offset 33KiB. Now starting with i.MX8MP the now
superfluous 1KiB offset was removed do that the actual image now starts
at 32KiB.
Unfortunately the 1KiB offset is woven into the offsets of the IVT
whereas the 32KiB are not, which means that we really have to handle
both offsets individually instead of just handling the sum of the
offsets.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>